+++ components/_standard/OrderedFieldsDataGrid.tsx
@@ -27,7 +27,6 @@ import Button from '@mui/material/Button';
 import AddIcon from '@mui/icons-material/Add';
 import DeleteIcon from '@mui/icons-material/Delete';
 import EditIcon from '@mui/icons-material/Edit';
-import SaveIcon from '@mui/icons-material/Save';
 
 interface OrderedFieldsDataGridProps {
   initialFields?: GridRowsProp;